requirements.txt
tensorflow >= 2.1.0
bert-for-tf2 >= 0.14.4
pandas
sklearn
nltk